+++ src/io/output.jl
@@ -35,6 +35,7 @@ function initialize_output(args::Dict{String,<:Any})::Dict{String,Any}
         "Runtime timestamp" => "$(Dates.now())",
         "Optimal switching metadata" => Dict{String,Any}[],
         "Optimal dispatch metadata" => Dict{String,Any}(),
+        "Fault studies metadata" => Dict{String,Any}[],
     )
 end
 
@@ -107,6 +108,7 @@ function analyze_results!(args::Dict{String,<:Any})::Dict{String,Any}
     get_timestep_stability!(args)
 
     get_timestep_fault_currents!(args)
+    get_timestep_fault_study_metadata!(args)
 
     return args["output_data"]
 end